ロジックの分離
