+++ src/service/reflectionService.ts
@@ -1,5 +1,6 @@
 import { reflectionRepository } from "../infrastructure/repository/reflectionRepository";
 import { toJST } from "../utils/date-helper";
+import prisma from "@/src/lib/prisma";
 
 const COUNT_PER_PAGE = 12;
 
@@ -21,6 +22,108 @@ export const reflectionService = {
     return {
       reflections,
       totalPage
+    };
+  },
+
+  async getByUsername(
+    page: number,
+    userId: string,
+    isCurrentUser: boolean,
+    tag?: string
+  ) {
+    const offset = (page - 1) * COUNT_PER_PAGE;
+    const tagFilter = tag && { [tag]: true };
+
+    const filteredReflectionCount = await prisma.reflection.count({
+      where: {
+        userId,
+        isPublic: isCurrentUser ? undefined : true,
+        ...tagFilter
+      }
+    });
+
+    const totalPage = Math.ceil(filteredReflectionCount / COUNT_PER_PAGE);
+
+    const userWithReflections = await prisma.user.findUnique({
+      where: {
+        id: userId
+      },
+      select: {
+        image: true,
+        bio: true,
+        goal: true,
+        website: true,
+        reflections: {
+          where: {
+            userId,
+            isPublic: isCurrentUser ? undefined : true,
+            ...tagFilter
+          },
+          orderBy: [{ isPinned: "desc" }, { createdAt: "desc" }],
+          take: COUNT_PER_PAGE,
+          skip: offset,
+          select: {
+            title: true,
+            reflectionCUID: true,
+            charStamp: true,
+            createdAt: true,
+            isPublic: true,
+            isPinned: true
+          }
+        }
+      }
+    });
+
+    // MEMO: タグ別の投稿数を全て取得しておく
+    const isLearningCount = await prisma.reflection.count({
+      where: {
+        userId,
+        isLearning: true,
+        isPublic: isCurrentUser ? undefined : true
+      }
+    });
+    const isAwarenessCount = await prisma.reflection.count({
+      where: {
+        userId,
+        isAwareness: true,
+        isPublic: isCurrentUser ? undefined : true
+      }
+    });
+    const isMonologueCount = await prisma.reflection.count({
+      where: {
+        userId,
+        isMonologue: true,
+        isPublic: isCurrentUser ? undefined : true
+      }
+    });
+    const isInputLogCount = await prisma.reflection.count({
+      where: {
+        userId,
+        isInputLog: true,
+        isPublic: isCurrentUser ? undefined : true
+      }
+    });
+    const isDailyReflectionCount = await prisma.reflection.count({
+      where: {
+        userId,
+        isDailyReflection: true,
+        isPublic: isCurrentUser ? undefined : true
+      }
+    });
+
+    const tagCountList = {
+      isDailyReflection: isDailyReflectionCount,
+      isLearning: isLearningCount,
+      isAwareness: isAwarenessCount,
+      isMonologue: isMonologueCount,
+      isInputLog: isInputLogCount
+    };
+
+    return {
+      userWithReflections,
+      totalPage,
+      filteredReflectionCount,
+      tagCountList
     };
   },
